Mega squash of these commits : Uber Optimization build: cleanup SDCLANG logic * Flags which are already passed to cflags don't need to be added to   cppflags and conlyflags since at the end, cflags get stacked on top   anyways.
* Also, unset SDCLANG_PRECONFIGURED_FLAGS since it is meant to be a
  local variable.

Change-Id: Id96e045fb7964aba733ff2e13c411d9adfd660aa
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

build: ensure QCLD usage when requested
* Passing -fuse-ld=qcld to cflags
  is required to make sure compiler
  chooses QCLD as linker.

Change-Id: Ifa7b5e40c51476e14a3b50465e2f5b58abeb8fee
Signed-off-by: Alex Naidis <alex.naidis@linux.com>

build: use the system's ccache by default
* ccache gets updated semi-frequently and some builders would like
  to be able to take advantage of new features without having to
  rely on the prebuilt version getting updated

* If a build system has ccache installed already, use that version
  instead of the prebuilt

Change-Id: I8988c8a25fab3694d84633f957c2b92ce84cf69e
Signed-off-by: Ícaro Hoff <icarohoff@gmail.com>

Also clean up our uber optimization script and change it's logic.
Use our own version of clang instead of google's snapshot.
Bring back strict aliasing, graphite on gcc, polly on clang and O3.
Disable optimization not working with sdclang if it is used.

Thanks @frap129 for the polly flags
Thanks @mydongistiny for the updated list of not working modules with polly.

I'll be cleaning up the list of not working modules later as it seems like there is a lot.

LTO and thin-lto was left out as i could not get my phone booting with these enabled.